Fixed bug caused by distance(int, int) signature on certain OMGraphics not  being called because they didn't override the distance(double, double) method of superclass.
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@2058 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/omGraphics/OMPoly.java
@@ -866,7 +866,7 @@ public class OMPoly extends OMAbstractLine implements Serializable {
      * @param y vertical pixel location.
      * @return the distance of the object to the location given.
      */
-    public float distance(int x, int y) {
+    public float distance(double x, double y) {
         if (shape != null) {
             return super.distance(x, y);
         }
